requirements.txt
torch
torchvision
transformers
opencv-python
Pillow
pytesseract
pdf2image
python-dotenv
scikit-learn
pandas
numpy
tqdm